split diagnostics into new module, review comments
