+++ pybatfish/client/commands.py
@@ -20,25 +20,20 @@ import base64
 import json
 import logging
 import os
-import shutil
 import tempfile
-from datetime import datetime
-from hashlib import md5
 from typing import Any, Dict, List, Optional, Union  # noqa: F401
 
-import requests
 import six
-from netconan import netconan
 from requests import HTTPError
 
 from pybatfish.client.consts import CoordConsts, WorkStatusCode
+from pybatfish.client.diagnostics import upload_diagnostics
 from pybatfish.datamodel.primitives import (  # noqa: F401
     Edge, Interface)
 from pybatfish.datamodel.referencelibrary import (NodeRoleDimension,
                                                   NodeRolesData, ReferenceBook,
                                                   ReferenceLibrary)
 from pybatfish.exception import BatfishException
-from pybatfish.question.question import QuestionBase
 from pybatfish.settings.issues import IssueConfig  # noqa: F401
 from pybatfish.util import (BfJsonEncoder, get_uuid, validate_name, zip_dir)
 from . import resthelper, restv2helper, workhelper
@@ -51,30 +46,6 @@ from .workhelper import (get_work_status,
 _bfDebug = True
 bf_logger = logging.getLogger("pybatfish.client")
 bf_session = Session(bf_logger)
-
-_INIT_INFO_QUESTIONS = [
-    QuestionBase({
-        "class": "org.batfish.question.initialization.ParseWarningQuestion",
-        "differential": False,
-        "instance": {
-            "instanceName": "__parseWarning",
-        }
-    }),
-    QuestionBase({
-        "class": "org.batfish.question.initialization.FileParseStatusQuestion",
-        "differential": False,
-        "instance": {
-            "instanceName": "__fileParseStatus",
-        }
-    }),
-    QuestionBase({
-        "class": "org.batfish.question.initialization.ConversionWarningQuestion",
-        "differential": False,
-        "instance": {
-            "instanceName": "__viConversionWarning",
-        }
-    }),
-]
 
 if _bfDebug:
     bf_logger.setLevel(logging.INFO)
@@ -719,16 +690,15 @@ def bf_set_snapshot(name=None, index=None):
     return bf_session.baseSnapshot
 
 
-def bf_upload_init_info(bucket, region, dry_run=True,
-                        netconan_config=None):
-    # type: (str, str, bool, str) -> str
+def bf_upload_diagnostics(dry_run=True, netconan_config=None):
+    # type: (bool, str, Optional[str], Optional[str]) -> str
     """
-    Fetch, anonymize, and optionally upload snapshot initialization information.
+    Fetch, anonymize, and optionally upload snapshot diagnostics information.
 
-    :param bucket: name of the AWS S3 bucket to upload to
-    :type bucket: string
-    :param region: name of the region containing the bucket
-    :type region: string
+    This runs a series of diagnostic questions on the current snapshot (including collecting parsing and conversion information).  The information collected is anonymized with Netconan which either anonymizes passwords and IP addresses (default) or uses the settings in the provided config.  The anonymous information is then either saved locally (if dry_run is True) or uploaded to an S3 bucket Batfish devs have access to (if dry_run is False).
+
+    This function is useful when
+
     :param dry_run: whether or not to skip upload; if False, anonymized files will be stored locally, otherwise anonymized files will be uploaded to the specified S3 bucket
     :type dry_run: bool
     :param netconan_config: path to Netconan configuration file
@@ -736,116 +706,7 @@ def bf_upload_init_info(bucket, region, dry_run=True,
     :return: location of anonymized files (local directory if doing dry run, otherwise S3 resource URL)
     :rtype: string
     """
-    tmp_dir = tempfile.mkdtemp()
-
-    for q in _INIT_INFO_QUESTIONS:
-        instance_name = q.get_name()
-        try:
-            content = json.dumps(q.answer().dict(), indent=4, sort_keys=True)
-        except BatfishException as e:
-            content = "Failed to answer {}: {}".format(instance_name, e)
-            bf_logger.warning(content)
-
-        with open(os.path.join(tmp_dir, instance_name), 'w') as f:
-            f.write(content)
-
-    tmp_dir_anon = tempfile.mkdtemp()
-    _anonymize_dir(tmp_dir, tmp_dir_anon, netconan_config)
-    shutil.rmtree(tmp_dir)
-
-    if dry_run:
-        bf_logger.info(
-            'See anonymized files produced by dry-run here: {}'.format(
-                tmp_dir_anon))
-        return tmp_dir_anon
-
-    if bucket is None:
-        raise ValueError('Bucket must be set to upload init info.')
-    if region is None:
-        raise ValueError('Region must be set to upload init info.')
-
-    # Generate anonymous S3 subdirectory name
-    anon_dir_name = md5(
-        '{}{}{}'.format(bf_session.network,
-                        bf_session.baseSnapshot,
-                        datetime.now()).encode()).hexdigest()
-    upload_dest = _generate_s3_url(bucket, region, anon_dir_name)
-
-    _upload_dir_to_url(upload_dest, tmp_dir_anon)
-    bf_logger.info('Uploaded files to: {}'.format(upload_dest))
-    shutil.rmtree(tmp_dir_anon)
-    return upload_dest
-
-
-def _anonymize_dir(input_dir, output_dir, netconan_config=None):
-    # type: (str, str, str) -> None
-    """
-    Anonymize files in input dir and save to output dir.
-
-    Uses Netconan with the provided configuration file to perform anonymization.  If no configuration is provided, only IP addresses and password will be anonymized.
-
-    :param input_dir: directory containing files to anonymize
-    :type input_dir: string
-    :param output_dir: directory to store anonymized files in
-    :type output_dir: string
-    :param netconan_config: path to Netconan configuration file
-    :type netconan_config: string
-    """
-    args = [
-        '-i', str(input_dir),
-        '-o', str(output_dir),
-    ]
-    if netconan_config is not None:
-        args.extend([
-            '-c', netconan_config
-        ])
-    else:
-        args.extend([
-            '-a',
-            '-p',
-        ])
-    netconan.main(args)
-
-
-def _generate_s3_url(bucket, region, resource):
-    # type: (str, str, str) -> str
-    """
-    Generate the S3 url for the specified resource.
-
-    :param bucket: name of the AWS S3 bucket to upload to
-    :type bucket: string
-    :param region: name of the AWS region containing the bucket
-    :type region: string
-    :param resource: name of the resource inside the bucket
-    :type resource: string
-    :return: URL for the specified resource
-    :rtype: string
-    """
-    return 'https://{bucket}.s3-{region}.amazonaws.com/{resource}'.format(
-        bucket=bucket, region=region, resource=resource)
-
-
-def _upload_dir_to_url(base_url, src_dir):
-    # type: (str, str) -> None
-    """
-    Recursively put files from the specified directory to the specified URL.
-
-    :param base_url: URL to put files to
-    :type base_url: string
-    :param src_dir: directory containing files to upload
-    :type src_dir: string
-    """
-    for root, dirs, files in os.walk(src_dir):
-        for name in files:
-            path = os.path.join(root, name)
-            rel_path = os.path.relpath(path, src_dir)
-            with open(path, 'rb') as data:
-                resource = '{}/{}'.format(base_url, rel_path)
-                r = requests.put(resource, data=data)
-                if r.status_code != 200:
-                    bf_logger.error(
-                        'Failed to upload resource: {} with status code {}'.format(
-                            resource, r.status_code))
+    return upload_diagnostics(dry_run=dry_run, netconan_config=netconan_config)
 
 
 def bf_write_question_settings(settings, question_class, json_path=None):
