+++ tests/integration/test_snapshot_funcs.py
@@ -18,7 +18,7 @@ import pytest
 import requests
 from requests import HTTPError
 
-from pybatfish.client.commands import (_INIT_INFO_QUESTIONS, bf_delete_network,
+from pybatfish.client.commands import (bf_delete_network,
                                        bf_delete_snapshot, bf_fork_snapshot,
                                        bf_generate_dataplane,
                                        bf_get_snapshot_inferred_node_role_dimension,
@@ -28,8 +28,9 @@ from pybatfish.client.commands import (_INIT_INFO_QUESTIONS, bf_delete_network,
                                        bf_init_snapshot, bf_list_snapshots,
                                        bf_put_node_roles,
                                        bf_session, bf_set_network,
-                                       bf_set_snapshot, bf_upload_init_info)
+                                       bf_set_snapshot, bf_upload_diagnostics)
 from pybatfish.client.consts import BfConsts
+from pybatfish.client.diagnostics import _INIT_INFO_QUESTIONS
 from pybatfish.client.extended import (bf_get_snapshot_input_object_text,
                                        bf_get_snapshot_object_text,
                                        bf_put_snapshot_object)
@@ -39,10 +40,6 @@ from pybatfish.datamodel.referencelibrary import (NodeRoleDimension,
 
 _this_dir = abspath(dirname(realpath(__file__)))
 _root_dir = abspath(join(_this_dir, pardir, pardir))
-
-# Public bucket to dump test files to
-_S3_BUCKET = "test-world-writeable-bucket"
-_S3_REGION = "us-west-2"
 
 
 @pytest.fixture()
@@ -246,9 +243,10 @@ def test_get_snapshot_object(network, example_snapshot):
     assert bf_get_snapshot_object_text('new_object') == 'goodbye'
 
 
-def test_upload_init_info(network, example_snapshot):
+def test_upload_diagnostics(network, example_snapshot):
     """Upload initialization information for example snapshot."""
-    resource_url = bf_upload_init_info(_S3_BUCKET, _S3_REGION, dry_run=False)
+    # Use public bucket to dump test files to
+    resource_url = bf_upload_diagnostics(dry_run=False)
 
     try:
         # Confirm files exist for each of the init info questions
@@ -259,6 +257,7 @@ def test_upload_init_info(network, example_snapshot):
         # Cleanup any successfully uploaded questions
         for q in _INIT_INFO_QUESTIONS:
             try:
-                requests.delete('{}/{}'.format(resource_url, q.get_name()))
+                pass
+                # requests.delete('{}/{}'.format(resource_url, q.get_name()))
             except Exception:
                 pass
